+++ R/auk-zerofill.r
@@ -23,13 +23,15 @@
 #'   as scientific or English common names, or a mixture of both. These names
 #'   must match the official eBird Taxomony ([ebird_taxonomy]). To include all
 #'   species, leave this argument blank.
-#' @param sep character; single character used to separate fields within a row.
-#' @param unique logical; should [auk_unique()] be run on the input data if it
-#'   hasn't already.
 #' @param collapse logical; whether to call `collapse_zerofill()` to return a
 #'   data frame rather than an `auk_zerofill` object.
+#' @param unique logical; should [auk_unique()] be run on the input data if it
+#'   hasn't already.
 #' @param complete logical; if `TRUE` (the default) all checklists are required 
 #'   to be complete prior to zero-filling.
+#' @param rollup logical; should [auk_rollup()] be run on the input data if it
+#'   hasn't already.
+#' @param sep character; single character used to separate fields within a row.
 #' @param ... additional arguments passed to methods.
 #'
 #' @details
@@ -70,8 +72,9 @@ auk_zerofill <- function(x, ...) {
 
 #' @export
 #' @describeIn auk_zerofill EBD data frame.
-auk_zerofill.data.frame <- function(x, sampling_events, species, unique = TRUE, 
-                                    collapse = FALSE, complete = TRUE, ...) {
+auk_zerofill.data.frame <- function(x, sampling_events, species, 
+                                    collapse = FALSE, unique = TRUE, 
+                                    complete = TRUE, rollup = TRUE, ...) {
   # checks
   assertthat::assert_that(
     is.data.frame(sampling_events),
@@ -125,7 +128,7 @@ auk_zerofill.data.frame <- function(x, sampling_events, species, unique = TRUE,
   }
   
   # check that auk_rollup has been run
-  if (!isTRUE(attr(x, "rollup"))) {
+  if (rollup && !isTRUE(attr(x, "rollup"))) {
     x <- auk_rollup(x)
   }
 
@@ -191,9 +194,10 @@ auk_zerofill.data.frame <- function(x, sampling_events, species, unique = TRUE,
 
 #' @export
 #' @describeIn auk_zerofill Filename of EBD.
-auk_zerofill.character <- function(x, sampling_events, species, sep = "\t",
-                                   unique = TRUE, collapse = FALSE,
-                                   complete = TRUE, ...) {
+auk_zerofill.character <- function(x, sampling_events, species, 
+                                   collapse = FALSE, unique = TRUE, 
+                                   complete = TRUE, rollup = TRUE, 
+                                   sep = "\t", ...) {
   # checks
   assertthat::assert_that(
     assertthat::is.string(x), file.exists(x),
@@ -202,20 +206,23 @@ auk_zerofill.character <- function(x, sampling_events, species, sep = "\t",
     assertthat::is.string(sep), nchar(sep) == 1, sep != " ")
 
   # read in the two files
-  ebd <- read_ebd(x = x, sep = sep, unique = FALSE)
+  ebd <- read_ebd(x = x, sep = sep, unique = FALSE, rollup = FALSE)
   sed <- read_sampling(x = sampling_events, sep = sep, unique = FALSE)
 
   # pass on to df method
-  auk_zerofill(x = ebd, sampling_events = sed, species = species,
-               unique = unique, collapse = collapse, complete = complete)
+  auk_zerofill(x = ebd, sampling_events = sed, species = species, 
+               collapse = collapse, unique = unique, complete = complete,
+               rollup = rollup)
 }
 
 #' @export
 #' @describeIn auk_zerofill `auk_ebd` object output from [auk_filter()]. Must
 #'   have had a sampling event data file set in the original call to
 #'   [auk_ebd()].
-auk_zerofill.auk_ebd <- function(x, species, sep = "\t", unique = TRUE, 
-                                 collapse = FALSE, complete = TRUE, ...) {
+auk_zerofill.auk_ebd <- function(x, species, 
+                                 collapse = FALSE, unique = TRUE, 
+                                 complete = TRUE, rollup = TRUE,
+                                 sep = "\t", ...) {
   # check that output files defined
   if (is.null(x$output)) {
     stop("No output EBD file in this auk_ebd object, try calling auk_filter().")
@@ -226,8 +233,9 @@ auk_zerofill.auk_ebd <- function(x, species, sep = "\t", unique = TRUE,
 
   # pass on to file method
   auk_zerofill(x = x$output, sampling_events = x$output_sampling,
-               species = species, sep = sep, unique = unique,
-               collapse = collapse, complete = complete)
+               species = species, collapse = collapse, 
+               unique = unique, complete = complete, rollup = rollup,
+               sep = sep)
 }
 
 #' @rdname auk_zerofill