+++ NEWS.md
@@ -9,7 +9,7 @@
 refer to the file name, rather than specifying the full path every time.
 - "Area" protocol added to `auk_protocol()` filter.
 - `auk_extent()` renamed `auk_bbox()`; `auk_extent()` deprecated and redirects to `auk_bbox()`
-- `auk_zerofill()` now throws an error if any checklists are not complete
+- `auk_zerofill()` now checks for complete checklists and gives option to not rollup
 - Fixed package load error when `EBD_PATH` is invalid
 - Fixed bug when reading files with a blank column using `readr`